delete jaar_config and redistribute functions
