+++ src/f02_bud/bud_config.py
@@ -1,0 +1,2 @@
+def max_tree_traverse_default() -> int:
+    return 20
