+++ src/f04_kick/test_kick/test_kick_file.py
@@ -1,6 +1,5 @@
 from src.f00_instrument.file import create_path, open_json
 from src.f01_road.deal import owner_name_str
-from src.f01_road.jaar_config import get_test_fisc_title as fisc_title
 from src.f04_kick.atom_config import face_name_str
 from src.f04_kick.delta import buddelta_shop
 from src.f04_kick.kick import kickunit_shop, create_kickunit_from_files
@@ -18,7 +17,7 @@ from os.path import exists as os_path_exists
 
 def test_KickUnit_save_atom_file_SavesCorrectFile(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_owner_dir = create_path(x_owners_dir, sue_str)
@@ -48,7 +47,7 @@ def test_KickUnit_save_atom_file_SavesCorrectFile(env_dir_setup_cleanup):
 
 def test_KickUnit_atom_file_exists_ReturnsObj(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_owner_dir = create_path(x_owners_dir, sue_str)
@@ -75,7 +74,7 @@ def test_KickUnit_atom_file_exists_ReturnsObj(env_dir_setup_cleanup):
 
 def test_KickUnit_open_atom_file_ReturnsObj(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_owner_dir = create_path(x_owners_dir, sue_str)
@@ -102,7 +101,7 @@ def test_KickUnit_open_atom_file_ReturnsObj(env_dir_setup_cleanup):
 
 def test_KickUnit_save_kick_file_SavesCorrectFile(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_kick_id = 2
@@ -138,7 +137,7 @@ def test_KickUnit_save_kick_file_SavesCorrectFile(env_dir_setup_cleanup):
 
 def test_KickUnit_kick_file_exists_ReturnsObj(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_owner_dir = create_path(x_owners_dir, sue_str)
@@ -164,7 +163,7 @@ def test_KickUnit_kick_file_exists_ReturnsObj(env_dir_setup_cleanup):
 
 def test_KickUnit_save_files_CorrectlySavesFiles(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_owner_dir = create_path(x_owners_dir, sue_str)
@@ -201,7 +200,7 @@ def test_KickUnit_save_files_CorrectlySavesFiles(env_dir_setup_cleanup):
 
 def test_KickUnit_create_buddelta_from_atom_files_SetsAttr(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_owner_dir = create_path(x_owners_dir, sue_str)
@@ -234,7 +233,7 @@ def test_KickUnit_create_buddelta_from_atom_files_SetsAttr(env_dir_setup_cleanup
 
 def test_create_kickunit_from_files_ReturnsObj(env_dir_setup_cleanup):
     # ESTABLISH
-    x_fisc_dir = create_path(fiscs_dir(), fisc_title())
+    x_fisc_dir = create_path(fiscs_dir(), "accord23")
     x_owners_dir = create_path(x_fisc_dir, "owners")
     sue_str = "Sue"
     sue_owner_dir = create_path(x_owners_dir, sue_str)
