+++ src/f06_listen/examples/example_listen_deltas.py
@@ -1,4 +1,3 @@
-from src.f01_road.jaar_config import get_fisc_title_if_None
 from src.f01_road.road import FiscTitle
 from src.f02_bud.bud_tool import bud_acctunit_str, bud_itemunit_str
 from src.f04_kick.atom_config import acct_name_str, parent_road_str, item_title_str
@@ -13,7 +12,8 @@ from src.f04_kick.delta import BudDelta, buddelta_shop
 
 
 def get_atom_example_itemunit_sports(fisc_title: FiscTitle = None) -> BudAtom:
-    fisc_title = get_fisc_title_if_None(fisc_title)
+    if not fisc_title:
+        fisc_title = "accord23"
     sports_str = "sports"
     x_dimen = bud_itemunit_str()
     insert_itemunit_budatom = budatom_shop(x_dimen, atom_insert())
