+++ src/f06_listen/hubunit.py
@@ -8,15 +8,10 @@ from src.f00_instrument.file import (
     set_dir,
     get_integer_filenames,
     get_max_file_number,
+    get_json_filename,
 )
 from src.f00_instrument.dict_toolbox import get_empty_set_if_None
 from src.f00_instrument.db_toolbox import sqlite_connection
-from src.f01_road.jaar_config import (
-    grades_folder,
-    get_rootpart_of_keep_dir,
-    treasury_filename,
-    get_json_filename,
-)
 from src.f01_road.finance import (
     default_fund_coin_if_None,
     validate_fund_pool,
@@ -61,7 +56,7 @@ from src.f04_kick.kick import (
     create_kickunit_from_files,
 )
 from src.f06_listen.basis_buds import get_default_plan
-from src.f06_listen.hub_path import create_gut_path, create_plan_path
+from src.f06_listen.hub_path import treasury_filename
 from src.f06_listen.hub_tool import (
     save_gut_file,
     open_gut_file,
@@ -117,7 +112,7 @@ def get_keep_jobs_dir(x_keep_dir: str) -> str:
 
 
 def get_keep_grades_dir(x_keep_dir: str) -> str:
-    return create_path(x_keep_dir, grades_folder())
+    return create_path(x_keep_dir, "grades")
 
 
 @dataclass
@@ -462,9 +457,6 @@ class HubUnit:
     def owner_filename(self, owner_name: OwnerName) -> str:
         return get_json_filename(owner_name)
 
-    def treasury_filename(self) -> str:
-        return treasury_filename()
-
     def treasury_db_path(self) -> str:
         return create_path(self.keep_dir(), treasury_filename())
 
@@ -634,7 +626,7 @@ def hubunit_shop(
 
 
 def get_keep_path(x_hubunit: HubUnit, x_road: TitleUnit) -> str:
-    keep_root = get_rootpart_of_keep_dir()
+    keep_root = "itemroot"
     x_road = rebuild_road(x_road, x_hubunit.fisc_title, keep_root)
     x_list = get_all_road_titles(x_road, x_hubunit.bridge)
     keep_sub_path = get_directory_path(x_list=[*x_list])
