+++ src/f06_listen/test_hubunit/test_hubunit__core.py
@@ -11,7 +11,6 @@ from src.f01_road.finance import (
     default_fund_coin_if_None,
     validate_fund_pool,
 )
-from src.f01_road.jaar_config import get_test_fisc_mstr_dir, get_rootpart_of_keep_dir
 from src.f06_listen.hub_path import create_deals_dir_path
 from src.f06_listen.hubunit import HubUnit, hubunit_shop, get_keep_path
 from src.f06_listen.examples.listen_env import (
@@ -120,7 +119,7 @@ def test_hubunit_shop_ReturnsObjWhenEmpty():
     usa_road = create_road(nation_road, usa_str)
     texas_str = "Texas"
     texas_road = create_road(usa_road, texas_str)
-    fisc_mstr_dir = get_test_fisc_mstr_dir()
+    fisc_mstr_dir = env_dir()
     accord23_str = "accord23"
 
     # WHEN
@@ -162,7 +161,6 @@ def test_hubunit_shop_ReturnsObjWhenEmpty():
     assert sue_hubunit.grade_path(bob_str) == x_grade_path
     treasury_filename = "treasury.db"
     x_treasury_file_path = create_path(sue_hubunit.keep_dir(), treasury_filename)
-    assert sue_hubunit.treasury_filename() == treasury_filename
     assert sue_hubunit.treasury_db_path() == x_treasury_file_path
 
 
@@ -189,7 +187,7 @@ def test_get_keep_path_ReturnsObj():
     dallas_str = "dallas"
     elpaso_str = "el paso"
     kern_str = "kern"
-    itemroot = get_rootpart_of_keep_dir()
+    itemroot = "itemroot"
     texas_road = create_road_from_titles([itemroot, texas_str])
     dallas_road = create_road_from_titles([itemroot, texas_str, dallas_str])
     elpaso_road = create_road_from_titles([itemroot, texas_str, elpaso_str])
@@ -202,7 +200,7 @@ def test_get_keep_path_ReturnsObj():
     kern_path = get_keep_path(sue_hubunit, kern_road)
 
     # THEN
-    itemroot_dir = create_path(sue_hubunit._keeps_dir, get_rootpart_of_keep_dir())
+    itemroot_dir = create_path(sue_hubunit._keeps_dir, itemroot)
     print(f"{kern_road=}")
     print(f"{itemroot_dir=}")
     assert texas_path == create_path(itemroot_dir, texas_str)
