+++ src/f06_listen/test_listen/test_agendas_duty_job.py
@@ -1,5 +1,3 @@
-from src.f00_instrument.file import delete_dir, save_file
-from src.f01_road.jaar_config import get_json_filename
 from src.f02_bud.item import itemunit_shop
 from src.f02_bud.bud import budunit_shop
 from src.f06_listen.hubunit import hubunit_shop
