+++ src/f08_fisc/fisc.py
@@ -11,7 +11,6 @@ from src.f00_instrument.dict_toolbox import (
     get_dict_from_json,
     get_json_from_dict,
 )
-from src.f01_road.jaar_config import get_test_fisc_mstr_dir
 from src.f01_road.finance import (
     default_respect_bit_if_None,
     filter_penny,
@@ -455,6 +454,10 @@ def _get_ote1_max_past_event_int(
     if past_timepoints := {tp for tp in event_timepoints if int(tp) <= deal_time}:
         max_past_timepoint = max(past_timepoints)
         return ote1_owner_dict.get(max_past_timepoint)
+
+
+def get_test_fisc_mstr_dir():
+    return "src/f08_fisc/examples/fisc_mstr"
 
 
 def fiscunit_shop(
