+++ src/f08_fisc/test_fisc_/test_fisc_.py
@@ -1,10 +1,9 @@
-from src.f00_instrument.file import create_path
+from src.f00_instrument.file import create_path, get_json_filename
 from src.f01_road.finance import (
     default_fund_coin_if_None,
     default_respect_bit_if_None,
     filter_penny,
 )
-from src.f01_road.jaar_config import get_json_filename
 from src.f01_road.road import default_bridge_if_None
 from src.f01_road.deal import tranbook_shop
 from src.f02_bud.healer import healerlink_shop
