+++ src/f09_pidgin/test_pidgin/test_map_acct.py
@@ -1,7 +1,7 @@
-from src.f01_road.jaar_config import default_unknown_word_if_None
 from src.f01_road.road import default_bridge_if_None
 from src.f01_road.deal import owner_name_str
 from src.f04_kick.atom_config import event_int_str, face_name_str
+from src.f09_pidgin.pidgin_config import default_unknown_word_if_None
 from src.f09_pidgin.pidgin_config import (
     otx_bridge_str,
     inx_bridge_str,
