+++ src/f09_pidgin/test_pidgin/test_map_title.py
@@ -1,4 +1,3 @@
-from src.f01_road.jaar_config import default_unknown_word_if_None
 from src.f01_road.road import default_bridge_if_None
 from src.f04_kick.atom_config import face_name_str, event_int_str
 from src.f09_pidgin.pidgin_config import (
@@ -6,6 +5,7 @@ from src.f09_pidgin.pidgin_config import (
     inx_bridge_str,
     unknown_word_str,
     otx2inx_str,
+    default_unknown_word_if_None,
 )
 from src.f09_pidgin.map import (
     TitleMap,
