+++ src/f09_pidgin/test_pidgin/test_pidginunit_dict.py
@@ -1,10 +1,10 @@
-from src.f01_road.jaar_config import default_unknown_word_if_None
 from src.f01_road.road import default_bridge_if_None
 from src.f04_kick.atom_config import face_name_str, event_int_str
 from src.f09_pidgin.pidgin_config import (
     otx_bridge_str,
     inx_bridge_str,
     unknown_word_str,
+    default_unknown_word_if_None,
 )
 from src.f09_pidgin.pidgin import (
     pidginunit_shop,
