+++ src/f11_etl/test_tran/test_event_pidgin_csvs2pidgin_jsons.py
@@ -6,10 +6,9 @@ from src.f00_instrument.file import (
     delete_dir,
 )
 from src.f01_road.road import default_bridge_if_None
-from src.f01_road.jaar_config import default_unknown_word_if_None
 from src.f04_kick.atom_config import type_RoadUnit_str
+from src.f09_pidgin.pidgin_config import pidgin_filename, default_unknown_word_if_None
 from src.f09_pidgin.pidgin import get_pidginunit_from_json
-from src.f09_pidgin.pidgin_config import pidgin_filename
 from src.f11_etl.transformers import (
     etl_event_pidgin_csvs_to_pidgin_json,
     etl_otz_event_pidgins_csvs_to_otz_pidgin_jsons,
